refactor: Create utilities folder and put utils there
