+++ src/linter/ui5Types/utils/utils.ts
